pyproject.toml
[project]
name = "celestrak-historical"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "click-datetime>=0.4.0",
    "gitpython>=3.1.45",
]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "pre-commit>=4.2.0",
    "ruff>=0.12.7",
    "ty>=0.0.1a16",
]
